+++ src/package/themes/default/components/select.ts
@@ -24,11 +24,11 @@ export default {
     optionActiveBackgroundColor: 'rgba(var(--vf-blue-600-rgb), 0.14)',
     optionActiveTextColor: 'var(--vf-text-color)',
     small: {
-        padding: '0.375rem 0.5rem',
-        fontSize: '0.875rem',
+        padding: 'var(--vf-sizes-sm-padding-y) var(--vf-sizes-sm-padding-x)',
+        fontSize: 'var(--vf-sizes-sm-font-size)',
     },
     large: {
-        padding: '0.625rem 0.875rem',
-        fontSize: '1.125rem',
+        padding: 'var(--vf-sizes-lg-padding-y) var(--vf-sizes-lg-padding-x)',
+        fontSize: 'var(--vf-sizes-lg-font-size)',
     },
 };